requirements.txt
requests
python-dotenv
PyYAML
pandas   # Will be needed for screener_logic
numpy    # Often a dependency of pandas
